dataflow/requirements.txt
apache-beam[gcp]
google-genai
google-cloud-bigquery
google-cloud-aiplatform
